I would gladly raise my voice in praise of women, only they won't let me raise my voice. 		-- Winkle

+++ phpbackend/backend/manage_db/delete_database.php
@@ -6,16 +6,29 @@ $db = new DbConnect($admin, $pass);
 
 $db_name = $_GET['db_name'];
 
-$sql = "DROP DATABASE $db_name";
+function safety ($db_name){
+  $restricted_names = ['information_schema', 'mysql', 'performance_schema', 'sys', 'wdi'];
+  if (in_array($db_name, $restricted_names)){
+    logger('if statement is true');
+    echo "$db_name can't be deleted!";
 
-$result = $db->conn->query($sql);
+  } else {
 
-if($result){
-  logger("dropped database");
-  echo "dropped database $db_name";
-  // do true stuff
-}
-  else{
-    logger($db->conn->error);
-    // do false stuff
+    logger('if statemnet is false');
+    global $db;
+    $sql = "DROP DATABASE $db_name";
+    $result = $db->conn->query($sql);
+
+    if($result){
+      logger("dropped database");
+      echo "dropped database $db_name";
+      // do true stuff
+    }
+      else{
+        logger($db->conn->error);
+        // do false stuff
+    }
   }
+}
+
+safety($db_name);
